requirements.txt
requests
feedparser
beautifulsoup4
lxml
openai
pyyaml
